==> Files and scripts/0 Download Images.ipynb <==
# Load libraries and functions

import os
import codecs
import pandas as pd
from argparse import ArgumentParser
from urllib import request
from tqdm import trange
from joblib import Parallel, delayed

#https://stackoverflow.com/questions/36445193/splitting-one-csv-into-multiple-files-in-python
def split(filehandler, delimiter=',', row_limit=20000,
          output_name_template='drive/My Drive/Data/book32-listing-split_%s.csv', output_path='.', keep_headers=True):
    import csv
    reader = csv.reader(filehandler, delimiter=delimiter)
    current_piece = 1
    current_out_path = os.path.join(
        output_path,
        output_name_template % current_piece
    )
    current_out_writer = csv.writer(open(current_out_path, 'w'), delimiter=delimiter)
    current_limit = row_limit
    if keep_headers:
        headers = reader.__next__()
        current_out_writer.writerow(headers)
    for i, row in enumerate(reader):
        if i + 1 > current_limit:
            current_piece += 1
            current_limit = row_limit * current_piece
            current_out_path = os.path.join(
                output_path,
                output_name_template % current_piece
            )
            current_out_writer = csv.writer(open(current_out_path, 'w'), delimiter=delimiter)
            if keep_headers:
                current_out_writer.writerow(headers)
        current_out_writer.writerow(row)

# Don't run more than once!
#split(open('drive/My Drive/Data/book32-listing.csv', mode='r', encoding='utf-8', errors='ignore'))

# Download the images in groups using exsiting code

header_names = ['Amazon ID (ASIN)', 'Filename', 'Image URL', 'Title', 'Author', 'Category ID',
                'Category']

def download_image(i):
    filename = csv.iloc[i]['Filename']
    category = csv.iloc[i]['Category']
    inner_output_dirpath = os.path.join("drive/My Drive/images", category)
    if not os.path.isdir(inner_output_dirpath):
        os.mkdir(inner_output_dirpath)
    output_filepath = os.path.join(inner_output_dirpath, filename)

    url = csv.iloc[i]['Image URL']
    if not os.path.isfile(output_filepath):
        downloaded_img = request.urlopen(url)
        f = open(output_filepath, mode='wb')
        f.write(downloaded_img.read())
        downloaded_img.close()
        f.close()

if not os.path.isdir("drive/My Drive/images"):
    os.makedirs("drive/My Drive/images")


with codecs.open('drive/My Drive/Data/book32-listing-split_1.csv', mode='r', encoding='utf-8', errors='ignore') as f:
    csv = pd.read_csv(f, delimiter=",", header=None, names=header_names)

print('[Download images into "{}"]'.format("images"))

Parallel(n_jobs=-1)(delayed(download_image)(i) for i in trange(len(csv)))

with codecs.open('drive/My Drive/Data/book32-listing-split_2.csv', mode='r', encoding='utf-8', errors='ignore') as f:
    csv = pd.read_csv(f, delimiter=",", header=None, names=header_names)

print('[Download images into "{}"]'.format("images"))

Parallel(n_jobs=-1)(delayed(download_image)(i) for i in trange(len(csv)))

with codecs.open('drive/My Drive/Data/book32-listing-split_3.csv', mode='r', encoding='utf-8', errors='ignore') as f:
    csv = pd.read_csv(f, delimiter=",", header=None, names=header_names)

print('[Download images into "{}"]'.format("images"))

Parallel(n_jobs=-1)(delayed(download_image)(i) for i in trange(len(csv)))

with codecs.open('drive/My Drive/Data/book32-listing-split_4.csv', mode='r', encoding='utf-8', errors='ignore') as f:
    csv = pd.read_csv(f, delimiter=",", header=None, names=header_names)

print('[Download images into "{}"]'.format("images"))

Parallel(n_jobs=-1)(delayed(download_image)(i) for i in trange(len(csv)))

with codecs.open('drive/My Drive/Data/book32-listing-split_5.csv', mode='r', encoding='utf-8', errors='ignore') as f:
    csv = pd.read_csv(f, delimiter=",", header=None, names=header_names)

print('[Download images into "{}"]'.format("images"))

Parallel(n_jobs=-1)(delayed(download_image)(i) for i in trange(len(csv)))

with codecs.open('drive/My Drive/Data/book32-listing-split_6.csv', mode='r', encoding='utf-8', errors='ignore') as f:
    csv = pd.read_csv(f, delimiter=",", header=None, names=header_names)

print('[Download images into "{}"]'.format("images"))

Parallel(n_jobs=-1)(delayed(download_image)(i) for i in trange(len(csv)))

with codecs.open('drive/My Drive/Data/book32-listing-split_7.csv', mode='r', encoding='utf-8', errors='ignore') as f:
    csv = pd.read_csv(f, delimiter=",", header=None, names=header_names)

print('[Download images into "{}"]'.format("images"))

Parallel(n_jobs=-1)(delayed(download_image)(i) for i in trange(len(csv)))

with codecs.open('drive/My Drive/Data/book32-listing-split_8.csv', mode='r', encoding='utf-8', errors='ignore') as f:
    csv = pd.read_csv(f, delimiter=",", header=None, names=header_names)

print('[Download images into "{}"]'.format("images"))

Parallel(n_jobs=-1)(delayed(download_image)(i) for i in trange(len(csv)))

with codecs.open('drive/My Drive/Data/book32-listing-split_9.csv', mode='r', encoding='utf-8', errors='ignore') as f:
    csv = pd.read_csv(f, delimiter=",", header=None, names=header_names)

print('[Download images into "{}"]'.format("images"))

Parallel(n_jobs=-1)(delayed(download_image)(i) for i in trange(len(csv)))

with codecs.open('drive/My Drive/Data/book32-listing-split_10.csv', mode='r', encoding='utf-8', errors='ignore') as f:
    csv = pd.read_csv(f, delimiter=",", header=None, names=header_names)

print('[Download images into "{}"]'.format("images"))

Parallel(n_jobs=-1)(delayed(download_image)(i) for i in trange(len(csv)))

with codecs.open('drive/My Drive/Data/book32-listing-split_11.csv', mode='r', encoding='utf-8', errors='ignore') as f:
    csv = pd.read_csv(f, delimiter=",", header=None, names=header_names)

print('[Download images into "{}"]'.format("images"))

Parallel(n_jobs=-1)(delayed(download_image)(i) for i in trange(len(csv)))
